// src/app/modules/logistica/entrega/coletas/models/coleta.ts
export interface ILogisticaEntregaColeta {
  ID_LOGI_FUSI_PEDI: number;
  CD_PEDI: string;
  DT_PEDI: string;
  QT_PESO: number;
  TIPO_PEDI: string;
  NM_CLIE: string;
  NM_FANT_CLIE: string;
  CD_CLIE: string;
  NOTA_FISC: number;
  DT_EMIS_NOTA_FISC: string;
  VL_NOT_FISC: number;
  DS_LOCA_ENTR: string;
  DS_BAIR: string;
  DS_CIDA: string;
  IBGE: number;
  DS_ESTA: string;
  CD_CEP: number;
  DS_PRAC: string;
  CD_PRAC: string;
  CD_REGI_ENTR: string;
  DS_REGI_ENTR: string;
  CD_EMPR: number;
  CD_DEPO: number;
  CD_EMPR_CNPJ_CPF: number;
  CD_CLIE_CNPJ_CPF: number;
  PRAZO: string;
  DS_OBSE?: string;
  DS_VEND: string;
  CD_REST: number;
  SEGM_CLIE: string;
  IN_PEDI_EXTE: string;
  DT_INCL: string;
  IN_STAT: string;
  CD_REFE: string;
  IN_STAT_ROMA_PEDI: string;
  ID_USUA_RESP_CADA: number;
  IN_INTE: string;
}